backend/requirements.txt
# Django
django>=5.0,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5

# Channels (WebSocket)
channels>=4.0
channels-redis>=4.1
daphne>=4.0

# Database
psycopg2-binary>=2.9

# Utils
python-dotenv>=1.0
reportlab>=4.0
Pillow>=10.0

# Production
gunicorn>=21.0
